[3.3.0.0] [BRK] Replace ChatHelper filters with Magic Chat Fraymework™ (Fixes #407)
IMPORTANT: [BREAKING] This commit removes all chat filtering methods and their respective configuration fields from ChatHelper. All code using that must be changed to use the new Magic Chat Fraymework™ located in the common.chat package. Deprecation of those symbols has been decided against because of limited usage and how difficult and pointless it would have been to keep them temporarily - Client code needs to migrate anyways.

This commit replaces all the filtering methods from ChatHelper with a new, improved chat framework. The framework is designed to be more object-oriented, easy to use, fluent to write and read and also individually configurable per instance. The previous implementation only had some public static (eeeew) fields to change filter behaviour globally. It fixes #407 and fixes #412.

Additionally, this commit adds unit tests for all new code, complying to JavaDoc.

The main purpose of this commit is to refactor and improve chat filtering code so that it can be reused easier and configured in a more fine-grained way. Also featured are improvements to the algorithm. Some functionality has been stripped because neither did it work well enough for a production environment nor did it provide much of a gain. All removed features have been disabled in Production because of countless false positives for ages.

Remarkable is the removal of the "dotlike chars" replacement, which would previously count [-_,~] as dots, but we have been unable to find a solution for the comma with an at least reasonable amount of false positives. I have never seen the other characters used in an actual advertisement, so that feature has been stripped entirely.