Add ResultSet deserialisation methods to XyLocation
This commit adds two methods to XyLocation, which can be used to extract XyLocations from a SQL ResultSet.

This also fixes a typo in a method name, which used to be 'pretyPrint', but has now been changed to 'prettyPrint'. The misspelled method has been deprecated.

(cherry picked from commit f85453c)