[BREAKING] Change project structure to reduce shading, Change package
This commit changes the project group id to li.l1t.common and adds
two new modules: xyc-games-bootstrap and xyc-bootstrap-legacy.

The former is the new standalone version of all libraries combined
with everything shaded. Individual libraries do not shade any more.

The latter is the same, but with the new li.l1t package remapped back
 to the legacy io.github.xxyy, so that existing projects can continue
 to use XYC without having to change all of their code. For now.

 Also, in the future, relocations will be reduced to the bare minimum.